Simplified data structures, refactored code, and fixed issue with getting the right contract when working with proxies
